refactoring - separating app and server for separation of concerns
