+++ src/server.ts
@@ -1,20 +1,7 @@
-import express from "express";
-import roverRouter from "../routes/roverRouter";
+import app from "../app";
 
-const app = express();
 const port = 3000;
-
-app.get("/", (req: any, res: any) => {
-  res.send(`Hello space traveller!
-    What are you exploring today?
-    A) /rover = all info
-    B) /rover/photos = photos from that rover`)
-});
-
-app.use("/rovers", roverRouter);
 
 app.listen(port, () => {
   console.log(`Test backend is running on port ${port}`);
 });
-
-export default app;
